requirements.txt
streamlit==1.44.1
python-dotenv==1.0.1
langchain==0.3.23
langchain-community==0.3.21
langchain-openai==0.3.12
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-chroma==0.2.2
chromadb==0.6.3
pypdf==5.4.0
docx2txt==0.8
